Extract validateAmount for method for reusability.
