+++ core/src/test/kotlin/org/stellar/anchor/util/SepHelperTest.kt
@@ -1,0 +1,15 @@
+package org.stellar.anchor.util
+
+import org.junit.jupiter.api.Test
+import org.stellar.sdk.xdr.MemoType
+
+internal class SepHelperTest {
+  @Test
+  fun test() {
+    assert(SepHelper.memoTypeString(MemoType.MEMO_ID).equals("id"))
+    assert(SepHelper.memoTypeString(MemoType.MEMO_HASH).equals("hash"))
+    assert(SepHelper.memoTypeString(MemoType.MEMO_TEXT).equals("text"))
+    assert(SepHelper.memoTypeString(MemoType.MEMO_NONE).equals("none"))
+    assert(SepHelper.memoTypeString(MemoType.MEMO_RETURN).equals("return"))
+  }
+}
